feat(85): fix name segementing
